+++ src/main/java/team18/team18_be/userInformation/repository/CompanyRepository.java
@@ -1,7 +1,7 @@
-package team18.team18_be.company.repository;
+package team18.team18_be.userInformation.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import team18.team18_be.company.entity.Company;
+import team18.team18_be.userInformation.entity.Company;
 
 public interface CompanyRepository extends JpaRepository<Company,Long> {
