+++ src/main/java/team18/team18_be/userInformation/repository/SignRepository.java
@@ -1,7 +1,7 @@
-package team18.team18_be.sign.Repository;
+package team18.team18_be.userInformation.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import team18.team18_be.sign.entity.Sign;
+import team18.team18_be.userInformation.entity.Sign;
 
 public interface SignRepository extends JpaRepository<Long, Sign> {
